Refactor #137 Separate employee CRUD processing and user account processing in separate files
